+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This list is not intended to be all-encompassing - it will document major and br
 changes with their rationale when appropriate. Given version `A.B.C.D`, breaking changes are to be expected in version number increments where changes in the `A` or `B` sections:
 
 ### v3.25.5.0
+- **http4k-connect-amazon-containercredentials** - Support passing of full CC URL and auth token header. This makes us compatible with AWS Snapstart
+-. 
 - **http4k-connect-*** - Upgrade dependencies.
 
 ### v3.25.4.0